style: add extract_word.c
